[3234] Extract flow/wizard class to file
